Overload methods, fix return value in literal initialisers
Add tests for literals that have values

+++ commit-based-cipm/fi-tests/cipm.consistency.fitests/src/cipm/consistency/fitests/similarity/java/initialiser/literals/IBinaryLongLiteralInitialiser.java
@@ -16,7 +16,7 @@ public interface IBinaryLongLiteralInitialiser extends ILongLiteralInitialiser {
     	return true;
     }
     
-    public default boolean setBinaryValue(BinaryLongLiteral bil, int val) {
+    public default boolean setBinaryValue(BinaryLongLiteral bil, long val) {
     	return this.setBinaryValue(bil, BigInteger.valueOf(val));
     }
 }
